requirements.txt
# Gmail API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# Configuration
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Machine Learning
sentence-transformers==3.3.1
torch==2.9.0
numpy==1.26.4
scikit-learn==1.5.2
lightgbm==4.5.0
pandas==2.2.3

# Visualization
matplotlib==3.9.2
seaborn==0.13.2

# Serialization
joblib==1.4.2

# Development/Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pre-commit==4.0.1
ruff==0.8.4
